% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rt.R
\name{align_rt}
\alias{align_rt}
\title{Align retention times to a reference data set}
\usage{
align_rt(x, model, var_name)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{infer_pf} function.}

\item{model}{A list containing the model output by the \code{loess} function.
The elements of this list are the models between the reference data set and
every other data set.}

\item{var_name}{A character string. The name of the retention time variable
that will be aligned with the reference data set (e.g., \verb{Feature apex}).}
}
\value{
A \code{data.table} with the retention times aligned to a reference
data set. The following variables have been added/removed:\tabular{ll}{
   Added \tab Removed \cr
   \code{RTalign} \tab  \cr
}
}
\description{
Aligns all retention times to a reference data set using the times from the
\code{Feature apex} column. The \code{loess} function from the \code{stats}
package is used for the alignment.
}
\author{
Evan A Martin
}
