% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_mass.R
\name{calc_error}
\alias{calc_error}
\title{Calculate ppm and retention time error}
\usage{
calc_error(x, ref_ds)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{align_rt} function.}

\item{ref_ds}{A character string specifying the name of the reference data
set.}
}
\value{
A list with three elements. The first element is the data matrix that
contains the error terms for the mass and retention times. These terms will
be used to both recalibrate the mass and cluster the data. The second
element is the median of the mass measurement error (in ppm) across all
data sets. The third element is the median of the standard deviation of the
retention time across all data sets.
}
\description{
Computes the error in ppm and retention time between each data set and a
reference data set. The default reference data set (returned by the
\code{find_ref_ds} function) is the one with the most unique proteoforms.
}
\author{
Evan A Martin
}
