% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_pcg.R
\name{collapse_pcg}
\alias{collapse_pcg}
\title{Choose a representative feature intensity}
\usage{
collapse_pcg(x, feature)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{gen_pcg} function. This
\code{data.table} contains the \code{pcGroup} variable which will be used
when choosing a representatvie feature intensity value.}

\item{feature}{A \code{data.table} containing the unidentified feature data
output from the \code{impute_fi} function.}
}
\value{
A \code{data.table} with the representative feature intensity values.
}
\description{
...
}
