% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine data sets}
\usage{
combine_data(x = list())
}
\arguments{
\item{x}{A list of data frames. All data frames must have the same number of
columns and the names of the columns must match exactly.}
}
\value{
...
}
\description{
Combines a group of data sets output from TopPIC. This function also adds
several variables that will be used throughout the top down workflow and
removes variables that are not used.
}
