% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_features.R
\name{combine_features}
\alias{combine_features}
\title{Combine identified and unidentified feature data}
\usage{
combine_features(ms2, ms1, summary_fn = "max")
}
\arguments{
\item{ms2}{A \code{data.table} containing the identified feature data.}

\item{ms1}{A \code{data.table} containing the unidentified feature data.}

\item{summary_fn}{A character string specifying the function that will be
used to select a representative feature intensity. The current functions
allowed are "max" or "sum". The default input is "max".}
}
\value{
A \code{data.table} containing the combined ms1 and ms2 feature
intensity values.
}
\description{
Combines the identified and unidentified feature data. The maximum feature
intensity is selected for each combination of Dataset, CV, Gene, and pcGroup
after joining the two data tables.
}
\author{
Vlad Petyuk
}
