% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_fdr.R
\name{compute_fdr}
\alias{compute_fdr}
\title{Compute the FDR}
\usage{
compute_fdr(x)
}
\arguments{
\item{x}{A \code{data.table} that contains decoy proteins.}
}
\value{
A named vector with the FDR at the PrSM, sequence, proteoform,
accession, and gene levels. The FDR values for each level are also printed
when the function is called.
}
\description{
Computes the FDR at the PrSM, sequence, proteoform, accession, and gene
levels.
}
\author{
Evan A Martin
}
