% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{create_gcc}
\alias{create_gcc}
\title{Combine clusters across genes}
\usage{
create_gcc(x, cutoff_ppm, cutoff_rt)
}
\arguments{
\item{x}{...}

\item{cutoff_ppm}{...}

\item{cutoff_rt}{...}
}
\value{
...
}
\description{
Combine two clusters if they have the same centroid (or are within the
specified mass/retention time threshold) but the genes of the two clusters
are different.
}
\author{
Evan A Martin
}
