% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_mdata}
\alias{create_mdata}
\title{Create feature metadata}
\usage{
create_mdata(x, cutoff_ppm, cutoff_rt)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{mk_pcg} function.}
}
\description{
This function uses the information from the identified feature data to create
a feature metadata data frame. This data frame will be used at the end of the
\code{TopPICR} pipeline when creating an \code{MSnSet} object.
}
\author{
Evan A Martin
}
