% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_msnset.R
\name{create_msnset}
\alias{create_msnset}
\title{Convert TopPICR output to an MSnSet object}
\usage{
create_msnset(
  x,
  meta,
  dataset_feature_data,
  dataset_sample_data,
  proteoform_def = c("Gene", "pcGroup"),
  partitioning = NULL
)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{match_features}
function.}

\item{meta}{A \code{data.table} output from the \code{create_mdata} function.}

\item{dataset_feature_data}{A \code{data.frame} optional argument
that links \code{Dataset} with fractions if sample were fractionated. This is
important because observation of proteoform in a fraction defined as separate
features.}

\item{dataset_sample_data}{A \code{data.frame} optional argument
that links \code{Dataset} with \code{sample_name}. Again, if the sample was
fractionated, multiple datsets correspond to a single sample. Additional
information in this table can be any other technical (e.g. batch, replicate)
or phenotypic information about samples.}

\item{proteoform_def}{A \code{character vector} describing how proteoform ID
is constructed. By default it is "Gene" and "pcGroup" (proteform cluster group).}

\item{partitioning}{A \code{character vector} describing how proteoform ID
is partitioned into features. For example, if the sample is fractionated then
\code{fraction} (defined in \code{dataset_feature_data} table) is part of the feature
definition. Also, if the dataset was acquired on an instrument with FAIMS
capability, then \code{CV} value (present in the first argument) is also part of
the feature definition. That is if the data was aquired with fractionation
and CV stepping then a proteoform is split into multiple features
(spread across fractions and CV values). Default is NULL, implying that there
was no partitioning of the proteoforms.}
}
\value{
An MSnSet object
}
\description{
Takes the combined data (output from the match between runs step),
dataset description and the proteoform metadata
and creates an MSnSet object.
}
\author{
Vlad Petyuk
}
