% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusters.R
\name{create_pcg}
\alias{create_pcg}
\title{Combines clusters based on mass and retention time}
\usage{
create_pcg(x, errors, n_mme_sd, n_Da, n_rt_sd)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{cluster} function.}

\item{errors}{A \code{list} output from the \code{calc_error} function. The
first element of the list contains the standard deviation and median of the
mass measurement error for each data set. The second element is the
standard deviation of the mass measurement error across all data sets. The
third element is the standard deviation of the retention time in seconds
across all data sets.}

\item{n_mme_sd}{The number of standard deviations used to create an envelope
around a cluster's centroid in mass space. The standard deviation is in
ppm. This threshold will be used to determine if two clusters will be
combined.}

\item{n_Da}{The number of Daltons used to create an envelope around a
cluster's centroid in mass space. Any cluster whose centroid falls within
this envelope will be included in the pool of clusters that could be
combined. This value is used to account for isotopic error and is separate
from \code{n_mme_sd}.}

\item{n_rt_sd}{The number of standard deviations used to create an envelope
around a cluster's centroid in retention time space. Any cluster whose
centroid falls within this envelope will be included in the pool of
clusters that could be combined.}
}
\value{
A \code{data.table} with the updated cluster assignment for each
gene/cluster combination.
}
\description{
Compares the centroid of each cluster within a gene and combines the two
clusters if the centroids fall within the specified threshold. The clusters
are combined by creating a new cluster membership variable \code{pcGroup}.
}
\author{
Evan A Martin
}
