% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{data_pf}
\alias{data_pf}
\title{...}
\usage{
data_pf(x, accession, ptm_annotation)
}
\arguments{
\item{x}{The x_meta data table (the output from create_mdata)}

\item{ptm_annotation}{A data frame that contains the mass of each PTM along
with its name. In the case when a PTM does not have a name the PTM's mass,
rounded to three decimal places, is used.}

\item{gene}{A character string indicating the accession to plot.}
}
\description{
...
}
