% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_to_msnset.R
\name{deprecated_to_msnset}
\alias{deprecated_to_msnset}
\title{Convert TopPICR output to an MSnSet object}
\usage{
deprecated_to_msnset(combined, meta)
}
\arguments{
\item{combined}{A \code{data.table} output from the \code{combine_features}
function.}

\item{meta}{A \code{data.table} output from the \code{create_mdata} function.}
}
\value{
An MSnSet object
}
\description{
Takes the combined data (output from the match between runs step) and the
meta data object and converts them to an MSnSet object.
}
\author{
Evan A Martin
}
