% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_toppic_version.R
\name{determine_toppic_version}
\alias{determine_toppic_version}
\title{A non-exported function. This function determines the version of the TopPIC
given the column names in the identification and features files.}
\usage{
determine_toppic_version(x)
}
\description{
A non-exported function. This function determines the version of the TopPIC
given the column names in the identification and features files.
}
