% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_accession.R
\name{filter_accession}
\alias{filter_accession}
\title{Filter UniProt accessions by count}
\usage{
filter_accession(x, count_by, threshold = 10)
}
\arguments{
\item{x}{...}

\item{count_by}{A Character string indicating what variable observations
should be counted by (e.g., Gene, ProteoForm, ...).}

\item{threshold}{An integer indicating the number of subjects (or project
IDs) a UniProt accession must occur in to be kept. The default is 10.}
}
\value{
...
}
\description{
Filters out UniProt accessions that occur in fewer than \code{threshold}
project IDs. Accessions are counted by the \code{count_by} argument. For
example, if \code{threshold} is 10 and \code{count_by} is Gene all rows
containing UniProt accessions that occur in fewer than 10 project IDs will be
removed. UniProt accessions are counted within Gene.
}
