% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pf.R
\name{filter_by_count}
\alias{filter_by_count}
\title{Filter low count proteoforms}
\usage{
filter_by_count(x, count_within, count, threshold)
}
\arguments{
\item{x}{A \code{data.table} output from \code{augment_annotation}.}

\item{count_within}{A character vector specifying the variables that will be
grouped before counting the number of occurrences of the \code{count}
variable.}

\item{count}{A character string indicating the variable that will be counted.}

\item{threshold}{An integer. The value that will be used to remove rows whose
count falls below the threshold.}
}
\value{
A \code{data.table} where all rows containing proteoforms with counts
below the threshold are filtered out.
}
\description{
Counts the number of samples each proteoform occurs in and filters out any
proteoforms that occur in fewer samples than the threshold.
}
