% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{find_collider}
\alias{find_collider}
\title{Find clusters across genes with same centroid}
\usage{
find_collider(x, cutoff_ppm, cutoff_rt)
}
\arguments{
\item{x}{...}

\item{cutoff_ppm}{...}

\item{cutoff_rt}{...}
}
\value{
...
}
\description{
Finds clusters clusters with the same centroid (or are within the
specified mass/retention time threshold) but the genes of the two clusters
are different and marks them as colliding genes.
}
\author{
Evan A Martin
}
