% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_fdr.R
\name{find_evalue_cutoff}
\alias{find_evalue_cutoff}
\title{Calculate E-value cutoff}
\usage{
find_evalue_cutoff(x, fdr_threshold)
}
\arguments{
\item{x}{A \code{data.table} output from either the \code{augment_annotation}
or \code{map_proteoform} function.}

\item{fdr_threshold}{A value between 0 and 1 indicating the desired FDR
level.}
}
\value{
A vector with three elements. These elements correspond to the
E-value cutoff for the three annotation types (SwissProt, VarSplic, and
TrEMBL). Note: the order of the annotation types depends on the order they
appear in the AnnType column.
}
\description{
Calculates the E-value cutoff that keeps the FDR under the given threshold
within each annotation type.
}
\author{
Evan A Martin
}
