% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rt.R
\name{find_ref_ds}
\alias{find_ref_ds}
\title{Find reference data set}
\usage{
find_ref_ds(x)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{infer_pf} function.}
}
\value{
A character vector containing the name of the reference data set. The
reference data set is the one that has the highest number of unique
\code{Proteoforms}.
}
\description{
\code{find_ref_ds} finds the data set that has the highest number of unique
proteoforms. This data set will be used as the reference data set when
aligning retention times and recalibrating the mass.
}
