% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rt.R
\name{form_model}
\alias{form_model}
\title{Use \code{loess} to create a model between data sets}
\usage{
form_model(x, ref_ds, ...)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{infer_pf} function.}

\item{ref_ds}{A character string specifying the reference data set.}

\item{...}{Arguments for the \code{\link[stats]{loess}} function.}
}
\value{
A list containing the loess model between each data set and the
reference data set.
}
\description{
Run the \code{\link[stats]{loess}} function using the values from the
\code{Feature apex} variable between the reference data set and all other
data sets.
}
\author{
Evan A Martin
}
