% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods_handling_fns.R
\name{get_mods_counts}
\alias{get_mods_counts}
\alias{get_distinct_mods}
\title{Report the modifications found on a protein}
\usage{
get_mods_counts(x, acc)

get_distinct_mods(x, accession)
}
\arguments{
\item{x}{A \code{data.frame} fData of an \code{MSnSet} object.}

\item{accession}{A character string specifying which UniProt accession will
be plotted.}
}
\description{
Report the modifications found on a protein
}
\section{Functions}{
\itemize{
\item \code{get_distinct_mods()}: 

}}
