% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_proteins.R
\name{infer_prot}
\alias{infer_prot}
\title{Protein inference}
\usage{
infer_prot(x)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{control_fdr} function.}
}
\value{
A \code{data.table} with all rows removed corresponding to the
UniProt accessions that do not belong to the set of accessions that map to
the largest number of \code{cleanSeq}s. The output from this function needs
to be kept for the rest of the \code{TopPICR} workflow as it will be used
by later functions.
}
\description{
The \code{infer_prot} function finds the smallest set of UniProt accessions
(\code{UniProtAcc}) that maps to the largest number of amino acid sequences
(\code{cleanSeq}).
}
\author{
Vlad Petyuk
}
