% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusters_oc.R
\name{merge_clusters_oc}
\alias{merge_clusters_oc}
\title{Merge clusters using the overlap coefficient}
\usage{
merge_clusters_oc(x, oc_cutoff)
}
\arguments{
\item{x}{...}

\item{oc_cutoff}{...}
}
\value{
...
}
\description{
...
}
