% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_accession_ptm}
\alias{plot_accession_ptm}
\title{...}
\usage{
plot_accession_ptm(
  x,
  accession,
  fill_by = "spectralCount",
  turbo = TRUE,
  viridis = NULL,
  name_from = c("Gene", "UniProtAcc"),
  mods_to_name = "all",
  border_color = "white",
  border_size = NULL,
  aa_step = 20,
  na_symbol = " ",
  save_plot = FALSE,
  file_path = NULL
)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{create_mdata} function.}

\item{accession}{...}

\item{fill_by}{...}

\item{name_from}{...}

\item{mods_to_name}{...}

\item{border_color}{...}

\item{border_size}{...}

\item{aa_step}{...}

\item{na_symbol}{...}

\item{save_plot}{...}

\item{file_path}{...}
}
\value{
...
}
\description{
...
}
\author{
Vlad Petyuk
}
