% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_accession_wgcna}
\alias{plot_accession_wgcna}
\title{Plotting proteoforms after WGCNA clustering}
\usage{
plot_accession_wgcna(
  x,
  accession,
  fill_by = "cluster",
  border_color = "black",
  ...
)
}
\arguments{
\item{x}{A \code{data.table} a typical use assumed \code{fData()} of an \code{MSnSet}.}

\item{accession}{A character string specifying which UniProt accession will
be plotted.}

\item{fill_by}{column name storing the assigned WGCNA clusters. According to
WGCNA convention those should be colors. Default values is "cluster".}

\item{border_color}{since sometimes WGCNA assigns colors close to white,
the default color for the border is "black".}

\item{...}{further arguments passed to \code{plot_accession_ptm()}}
}
\value{
A \code{ggplot2} object
}
\description{
Colors the proteoforms according to colors assigned by WGCNA clustering.
Derivative of the \code{plot_accession_ptm()} function.
}
\author{
Vlad Petyuk
}
