% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_fragments}
\alias{plot_fragments}
\title{Plot all fragments for a given UniProt accession}
\usage{
plot_fragments(
  x,
  gene = NULL,
  accession = NULL,
  turbo = TRUE,
  viridis = NULL,
  save_plot = FALSE,
  file_path = NULL
)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{create_mdata} function.}

\item{gene}{A character string. A plot will be produced for each UniProt
accession mapping to the specified gene. When a \code{gene} is specified
\code{accession} does not also need to be specified.}

\item{accession}{A character string specifying which UniProt accession will
be plotted.}

\item{turbo}{A logical value indicating whether the Turbo colormap will be
used.}

\item{viridis}{A character string indicating which colormap option will be
used. The available options are: "magma" (or "A"), "inferno" (or "B"),
"plasma" (or "C"), "viridis" (or "D") and "cividis" (or "E").}

\item{save_plot}{A logical value. If TRUE the plot will be saved as a PNG
file. When a plot is saved to a file it is not also displayed in R/RStudio.}

\item{file_path}{The path to the folder where each plot will be saved. This
must be specified if \code{save_plot} is TRUE.}
}
\value{
A plot of the fragments for a given accession. When \code{save_plot}
is TRUE a PNG file is created for each plot.
}
\description{
Plots all fragments associated with a given accession. If a gene name is
provided a separate plot will be produced for each accession within the
specified gene.
}
\author{
Evan A Martin
}
