% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_toppic_DMS.R
\name{read_TopPIC_DMS}
\alias{read_TopPIC_DMS}
\title{Read in TopPIC results from DMS.}
\usage{
read_TopPIC_DMS(data_package_num)
}
\arguments{
\item{data_package_num}{DMS data package number that contain TopPIC search
results. Note, it may not work correctly if multipe TopPIC searches
were done with different parameter or FASTA files.}
}
\value{
A list with MS/MS identification and MS features.
}
\description{
Works only within PNNL's DMS.
}
\author{
Vlad Petyuk
}
