% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_mass.R
\name{recalibrate_mass}
\alias{recalibrate_mass}
\title{Recalibrate the mass}
\usage{
recalibrate_mass(x, ref_ds)
}
\arguments{
\item{x}{The \code{data.table} output from the \code{align_rt} function.}

\item{ref_ds}{A character string specifying the name of the reference data
set.}
}
\value{
A list with three elements. The first element is the data matrix with
the recalibrated mass. The second element is the MAD of the error in mass
between the reference and non-reference data sets. This is reported in ppm.
The third element is the MAD of the retention time error between the
reference data set and all other data sets.
}
\description{
Calibrates the mass to a reference data set and calculates the error between
each data set and the reference data set in parts per million (ppm). The
median absolute deviation (MAD) of the ppm error is also calculated.
}
