% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_fdr.R
\name{remove_decoys}
\alias{remove_decoys}
\title{Remove decoy proteins}
\usage{
remove_decoys(x)
}
\arguments{
\item{x}{A \code{data.table} that contains decoy proteins. This function
should be called after \code{apply_evalue_cutoff} and before the
\code{align_rt} function.}
}
\value{
A \code{data.table} with the rows containing decoy proteins and the
\code{isDecoy} column removed.
}
\description{
This function filters out the decoy proteins and also removes the
\code{isDecoy} variable.
}
\author{
Evan A Martin
}
