% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mods_handling_fns.R
\name{retain_one_mod}
\alias{retain_one_mod}
\alias{remove_a_mod}
\title{Functions for removing modifications}
\usage{
retain_one_mod(x, mod_name)

remove_a_mod(x, mod_name)
}
\arguments{
\item{x}{A \code{data.frame} fData of an \code{MSnSet} object.}

\item{mod_name}{A character string specifying the modification to retain}
}
\description{
This type of preprocessing makes sense prior plotting
}
\section{Functions}{
\itemize{
\item \code{remove_a_mod()}: 

}}
