% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoform_level.R
\name{set_pf_level}
\alias{set_pf_level}
\title{Determine the proteoform level for each \code{Proteoform}}
\usage{
set_pf_level(x)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{augment_annotation} (or
another downstream) function.}
}
\value{
A \code{data.table} with the \code{Proteoform Level} variable added.
}
\description{
Assigns each \code{Proteoform} a level based on the classification system
proposed by Smith et al. 2019.
}
\references{
Smith, L.M., Thomas, P.M., Shortreed, M.R. et al. A five-level
classification system for proteoform identifications. Nat Methods 16,
939–940 (2019). https://doi.org/10.1038/s41592-019-0573-x
}
\author{
James Fulcher
}
