% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rt.R
\name{align_rt}
\alias{align_rt}
\title{...}
\usage{
align_rt(x, ref_ds, ...)
}
\arguments{
\item{x}{...}

\item{ref_ds}{A character string specifying the reference data set.}

\item{...}{Arguments for the \code{\link[stats]{loess}} function.}
}
\value{
...
}
\description{
...
}
