% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_annotation.R
\name{augment_annotation}
\alias{augment_annotation}
\title{Augment}
\usage{
augment_annotation(
  x,
  fst_path_norm,
  fst_path_decoy,
  fst_file_norm,
  fst_file_decoy
)
}
\arguments{
\item{x}{...}

\item{fst_path_norm}{...}

\item{fst_path_decoy}{...}

\item{fst_file_norm}{...}

\item{fst_file_decoy}{...}
}
\value{
...
}
\description{
...
}
