% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_pcg.R
\name{choose_fi}
\alias{choose_fi}
\title{Choose feature intensity from the best CV/Fraction}
\usage{
choose_fi(x)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{collapse_pcg} function.}
}
\value{
A \code{data.table} ...
}
\description{
...
}
\author{
Evan A Martin, Vlad Petyuk
}
