% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{...}
\usage{
cluster(x, method, height, min_size)
}
\arguments{
\item{x}{A list with three elements output from the x_47 function. The first
element is the the data matrix with the aligned retention times and
recalibrated mass. The second and third elements are the median of the ppm
and retention time errors respectively. These values are calculated by
Dataset.}

\item{method}{A character string indicating what agglomeration method should
be used in the hclust function.}

\item{height}{An number specifying the height at which the tree created by
the hclust function should be cut.}

\item{min_size}{An integer value indicating the minimum number of points a
cluster must have. All clusters with fewer members than min_size will be
reclassified as "noise" points.}
}
\value{
...
}
\description{
...
}
