% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_pcg.R
\name{combine_features}
\alias{combine_features}
\title{Combine identified and unidentified feature data}
\usage{
combine_features(x, feature)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{gen_pcg} function.}

\item{feature}{A \code{data.table} containing the unidentified feature data
output from the \code{impute_fi} function.}
}
\value{
A \code{data.table} ...
}
\description{
Combines the identified and unidentified feature data. The maximum feature
intensity is selected for each combination of Dataset, CV, Gene, and pcGroup
after joining the two data tables.
}
\author{
Vlad Petyuk
}
