% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_fdr.R
\name{control_fdr}
\alias{control_fdr}
\title{Filter the data by p-value}
\usage{
control_fdr(x, pvals)
}
\arguments{
\item{x}{...}

\item{pvals}{A vector of three numeric values to be used as p-value cutoffs.
The p-value cutoffs must be in the order of unique(x$AnnType). For example,
if unique(x$AnnType) = "SwissProt", "TrEMBL", "VarSplic" then the p-value
cutoff for SwissProt must be the first element of pvals, the p-value cutoff
for TrEMBL must be the second element of pvals, and the p-value cutoff for
VarSplic must be the third element of pvals.}
}
\value{
...
}
\description{
The \code{control_fdr()} function filters the data by p-value for each
annotation type. If the p-values from the \code{pval_cutoff()} function are
used the overall FDR will fall below the given threshold.
}
