% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_mdata}
\alias{create_mdata}
\title{Create feature metadata}
\usage{
create_mdata(x, errors, n_mme_sd, n_rt_sd)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{create_pcg} function.}

\item{errors}{A \code{list} output from the \code{calc_error} function. The
first element of the list contains the standard deviation and median of the
mass measurement error for each data set. The second element is the
standard deviation of the mass measurement error across all data sets. The
third element is the standard deviation of the retention time in seconds
across all data sets.}

\item{n_mme_sd}{A numeric value indicating the number of standard deviations
in ppm. This value is used as the threshold for determining if two clusters
(from two different genes) have the same centroid in mass space.}

\item{n_rt_sd}{A numeric value representing the number of standard
deviations in seconds. This value is used as the threshold for determining
if two clusters (from two different genes) have the same centroid in
retention time space.}
}
\value{
A \code{data.table} with a row for each unique combination of \code{Gene},
\code{pcGroup}, and \code{Proteoform}. The \code{collision} variable indicates which
clusters have the same centroid across different genes. The
\code{data.table} also contains other proteoform information such as
protein length, UniProt accession, and first and last amino acid among
others.
}
\description{
This function uses the information from the identified feature data to create
a feature metadata data frame. This data frame will be used at the end of the
\code{TopPICR} pipeline when creating an \code{MSnSet} object.
}
\author{
Evan A Martin
}
