% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_clusters.R
\name{create_pcg}
\alias{create_pcg}
\title{Combines clusters based on mass and retention time}
\usage{
create_pcg(x, errors, ppm_cutoff, n_Da, n_rt_sd)
}
\arguments{
\item{x}{...}

\item{errors}{...}

\item{ppm_cutoff}{...}

\item{n_Da}{...}

\item{n_rt_sd}{...}
}
\description{
Compares the centroid of each cluster within a gene and combines the two
clusters if the centroids fall within the specified threshold. The clusters
are combined by creating a new cluster membership variable \code{pcGroup}.
}
\author{
Evan A Martin
}
