% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusters.R
\name{create_pcg}
\alias{create_pcg}
\title{Combines clusters based on mass and retention time}
\usage{
create_pcg(x, errors, ppm_cutoff, n_Da, n_rt_sd)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{cluster} function.}

\item{errors}{A \code{list} output from the \code{calc_error} function.}

\item{ppm_cutoff}{The threshold in ppm that will be used to determine if two
clusters will be combined.}

\item{n_Da}{The number of Daltons used to create an envelope around a
cluster's centroid in mass space. Any cluster whose centroid falls within
this envelope will be included in the pool of clusters that could be
combined.}

\item{n_rt_sd}{The number of standard deviations used to create an envelope
around a cluster's centroid in retention time space. Any cluster whose
centroid falls within this envelope will be included in the pool of
clusters that could be combined.}
}
\value{
A \code{data.table} with the updated cluster assignment for each
gene/cluster combination. The following variables have been added/removed:\tabular{ll}{
   Added \tab Removed \cr
   \code{pcGroup} \tab  \cr
}
}
\description{
Compares the centroid of each cluster within a gene and combines the two
clusters if the centroids fall within the specified threshold. The clusters
are combined by creating a new cluster membership variable \code{pcGroup}.
}
\author{
Evan A Martin
}
