% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_rt.R
\name{form_model}
\alias{form_model}
\title{Use \code{loess} to create a model between data sets}
\usage{
form_model(x, ref_ds, ...)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{infer_prot} function.}

\item{ref_ds}{A character string specifying the reference data set.}

\item{...}{Arguments for the \code{\link[stats]{loess}} function.}
}
\value{
A list containing the loess model between each data set and the
reference data set.
}
\description{
Runs the \code{\link[stats]{loess}} function using the values from the
\code{Feature apex} variable between the reference data set and each of the
remaining data sets (one at a time). We refer to the data set that is being
aligned to the reference as the current or working data set. The model is
created with the proteoforms that are found in both the reference data set
and the working data set.
}
\author{
Evan A Martin
}
