% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_clusters.R
\name{gen_pcg}
\alias{gen_pcg}
\title{...}
\usage{
gen_pcg(x, errors, ppm_cutoff, n_Da, n_rt_sd)
}
\arguments{
\item{x}{...}

\item{errors}{...}

\item{ppm_cutoff}{...}

\item{n_Da}{...}

\item{n_rt_sd}{...}
}
\description{
...
}
\author{
Evan A Martin
}
