% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_features.R
\name{match_features}
\alias{match_features}
\title{Extract feature intensity values from unidentified feature data}
\usage{
match_features(
  ms2,
  ms1,
  errors,
  n_mme_sd,
  n_rt_sd,
  summary_fn = "max",
  cores = NULL
)
}
\arguments{
\item{ms2}{A \code{data.table} containing the identified feature data.}

\item{ms1}{A \code{data.table} containing the unidentified feature data.
This data is created from the files output by TopPIC ending with
"ms1.feature" and must contain the \code{RTalign} and \code{RecalMass}
variables. These variables are created with the unidentified feature data
in the \code{align_rt} and \code{recalibrate_mass} functions.}

\item{errors}{A \code{list} output from the \code{calc_error} function. The
first element of the list contains the standard deviation and median of the
mass measurement error for each data set. The second element is the
standard deviation of the mass measurement error across all data sets. The
third element is the standard deviation of the retention time in seconds
across all data sets.}

\item{n_mme_sd}{A numeric value indicating the number of standard
deviations to use when creating a cutoff in the mass dimension. This
threshold is used for determining whether an unidentified feature is close
enough in mass to be considered part of an identified feature cluster. The
mean \code{RecalMass} of all points in the cluster is used for the
comparison.}

\item{n_rt_sd}{A numeric value representing the number of standard deviations
to use when creating a retention time cutoff. This value is the threshold
used to determine if an unidentified feature is close enough in retention
time to be considered part of an identified feature cluster. The mean
retention time of all points in the cluster is used for the comparison.}

\item{summary_fn}{A character string specifying the function to use when
summarizing the feature intensity. The function must contain the
\code{na.rm} argument. Some examples of functions that are allowed are
\code{max}, \code{sum}, or \code{median}.}

\item{cores}{\code{integer} Number of cores for parallelization.
Default is NULL. If NULL then the maximum number of cores is used,
but no more than R's limit of 125.}
}
\value{
A \code{data.table} containing all unidentified features that fall
within the threshold of an identified feature gene/cluster combination.
}
\description{
Matches unidentified feature intensities with identified feature clusters.
The mass and retention time of each unidentified feature is compared to the
centroid of each identified cluster. If the unidentified feature falls within
the specified mass and retention time threshold it is added to the cluster.
}
\author{
Evan A Martin
}
