% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_features.R
\name{match_features}
\alias{match_features}
\title{Extract feature intensity values from unidentified feature data}
\usage{
match_features(ms2, ms1, errors, n_ppm_sd, n_rt_sd, summary_fn = "max")
}
\arguments{
\item{ms2}{A \code{data.table} containing the identified feature data.}

\item{ms1}{A \code{data.table} containing the unidentified feature data.
This data is created from the files output by TopPIC ending with
"ms1.feature" and must contain the \code{RTalign} and \code{RecalMass}
variables. These variables are created with the unidentified feature data
in the \code{align_rt} and \code{recalibrate_mass} functions.}

\item{errors}{A \code{list} output from the \code{calc_error} function. The
ppm and retention time standard deviations (across all data sets) are the
last two elements of this list. They will be used along with the
\code{n_ppm_sd} and \code{n_rt_sd} arguments for determining whether or not
an unidentified feature belongs to a given cluster.}

\item{n_ppm_sd}{The number of standard deviations used to create a ppm cutoff
when determining if an unidentified feature belongs to a given cluster. A
feature belongs to a cluster if its recalibrated mass and aligned retention
time fall within the mass/retention time envelope around the cluster's
centroid.}

\item{n_rt_sd}{The number of standard deviations used to create a retention
time cutoff when determining if an unidentified feature belongs to a given
cluster. A feature belongs to a cluster if its recalibrated mass and
aligned retention time fall within the mass/retention time envelope around
the cluster's centroid.}

\item{summary_fn}{A character string specifying the function to use when
summarizing the feature intensity. The function must contain the
\code{na.rm} argument. Some examples of functions that are allowed are
\code{max}, \code{sum}, or \code{median}.}
}
\value{
A \code{data.table} containing all unidentified features that fall
within the threshold of an identified feature gene/cluster combination.
}
\description{
Matches unidentified feature intensities with identified feature clusters.
The mass and retention time of each unidentified feature is compared to the
centroid of each identified cluster. If the unidentified feature falls within
the specified ppm threshold it is added to the cluster.
}
\author{
Evan A Martin
}
