% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_features.R
\name{match_features}
\alias{match_features}
\title{Extract feature intensity values from unidentified feature data}
\usage{
match_features(x, ms1, ppm_cutoff, rt_cutoff)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{merge_clusters} function.}

\item{ms1}{A \code{data.table} containing the unidentified feature data.
This data is created from the files output by TopPIC ending with
"ms1.feature" and must contain the \code{RTalign} and \code{RecalMass}
variables. These variables are created with the unidentified feature data
in the \code{align_rt} and \code{recalibrate_mass} functions.}

\item{ppm_cutoff}{An integer representing the retention time cutoff in
seconds. This value is the threshold used to determine if an unidentified
feature is close enough in retention time to be considered part of an
identified feature cluster. The mean retention time of all points in the
cluster is used for the comparison.}

\item{rt_cutoff}{An integer indicating the mass cutoff in ppm. This value is
used as the threshold for determining if an unidentified feature is close
enough in mass to be considered part of an identified feature cluster. The
mean \code{RecalMass} of all points in the cluster is used for the
comparison.}
}
\value{
A \code{data.table} containing all unidentified features that fall
within the threshold of an identified feature gene/cluster combination.
}
\description{
...
}
\author{
Evan A Martin
}
