% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusters.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Merge clusters using a mass/retention time envelope}
\usage{
merge_clusters(x, errors, min_size, ppm_cutoff, n_Da, n_rt_sd)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{cluster} function.}

\item{errors}{A \code{list} output from the \code{calc_error} function.}

\item{min_size}{An integer value indicating the minimum number of points a
cluster must have. All clusters with fewer members than min_size will be
reclassified as "noise" points.}

\item{ppm_cutoff}{The threshold used to determine if a point is close enough
in mass to be merged with a new cluster.}

\item{n_Da}{The number of Daltons (multiplied by the difference in mass
between Carbon 12 and Carbon 13) to use when creating the mass portion of
the mass/retention time envelope.}

\item{n_rt_sd}{The number of standard deviations to use when creating the
retention time portion of the mass/retention time envelope.}
}
\value{
A \code{data.table} with like clusters combined if the points in the
clusters fall within the mass/retention time envelope of the larger
cluster. The following variables have been added/removed:\tabular{ll}{
   Added \tab Removed \cr
   \code{cluster_new} \tab  \cr
}
}
\description{
Merges points with clusters based on their mass and retention time. Starting
with the largest cluster, each point is compared to a mass/retention time
envelope around the centroid of the cluster. If a point falls within this
envelope the point is merged with the cluster. This process is repeated for
each cluster (except the noise cluster).
}
\author{
Evan A Martin
}
