% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusters.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Merge clusters using at mass/retention time envelope}
\usage{
merge_clusters(x, min_size, ppm_cutoff, ppm_range, rt_range)
}
\arguments{
\item{x}{...}

\item{min_size}{...}

\item{ppm_cutoff}{...}

\item{ppm_range}{...}

\item{rt_range}{...}
}
\value{
...
}
\description{
...
}
