% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_fdr.R
\name{pval_cutoff}
\alias{pval_cutoff}
\title{Calculate p-value cutoff}
\usage{
pval_cutoff(x, fdr_threshold)
}
\arguments{
\item{x}{...}

\item{fdr_threshold}{...}
}
\value{
...
}
\description{
Calculates the p-value cutoff that keeps the FDR under the given threshold
within each annotation type.
}
