% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_toppic.R
\name{read_toppic}
\alias{read_toppic}
\title{Read in TopPIC files}
\usage{
read_toppic(file_path, file_name, ...)
}
\arguments{
\item{file_path}{...}

\item{file_name}{...}

\item{...}{Additional arguments to the \code{read.delim} function.}
}
\description{
...
}
