% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_mass.R
\name{recalibrate_mass}
\alias{recalibrate_mass}
\title{Recalibrate the mass}
\usage{
recalibrate_mass(x, errors, var_name)
}
\arguments{
\item{x}{The \code{data.table} output from the \code{align_rt} function.}

\item{errors}{A \code{list} output from the \code{calc_error} function. The
first element of the list contains the standard deviation and median of the
mass measurement error for each data set. The second element is the
standard deviation of the mass measurement error across all data sets. The
third element is the standard deviation of the retention time in seconds
across all data sets.}

\item{var_name}{A character string. The name of the mass variable (e.g.,
\verb{Precursor mass}) that will be recalibrated with the ppm error calculated
in the \code{calc_error} function.}
}
\value{
A \code{data.table} containing the recalibrated mass variable.
}
\description{
Recalibrates the mass using the error terms returned by the \code{calc_error}
function.
}
\author{
Evan A Martin
}
