% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_mass.R
\name{recalibrate_mass}
\alias{recalibrate_mass}
\title{Recalibrate the mass}
\usage{
recalibrate_mass(x, errors)
}
\arguments{
\item{x}{The \code{data.table} output from the \code{align_rt} function.}

\item{errors}{A list output from the \code{calc_error} function.}
}
\value{
A \code{data.table} containing the recalibrated mass variable. The
following variables were added/removed:\tabular{ll}{
   Added \tab Removed \cr
   \code{RecalMass} \tab  \cr
}
}
\description{
Calibrates the mass to a reference data set using the error terms returned by
the \code{calc_error} function.
}
