% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_mass.R
\name{recalibrate_mass}
\alias{recalibrate_mass}
\title{Recalibrate the mass}
\usage{
recalibrate_mass(x, x_align, ref_ds)
}
\arguments{
\item{x}{The data frame that was used as the input to the align_rt function.
This data frame is required because the align_rt function deletes columns
that are needed in this function and other downstream functions.}

\item{x_align}{A data frame output from the \code{align_rt} function.}

\item{ref_ds}{A character string specifying the reference data set.}
}
\value{
...
}
\description{
...
}
