% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad_genes.R
\name{rm_false_gene}
\alias{rm_false_gene}
\title{Select one gene per feature}
\usage{
rm_false_gene(x)
}
\arguments{
\item{x}{A \code{data.table} output from the \code{read_toppic} function.}
}
\value{
A \code{data.table}.
}
\description{
A given feature can occur in multiple genes. This function selects a single
gene for each feature. The gene with the minimum E-value is kept and all
other genes are removed.
}
\author{
Vlad Petyuk
}
